CurrencyRate entity added to persist data from api nbp

+++ src/main/java/com/nbp/currencyapp/dto/RateDTO.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Getter
@@ -15,6 +16,6 @@ public class RateDTO implements Serializable {
 
     private String currency;
     private String code;
-    private Double mid;
+    private BigDecimal mid;
 
 }
